+++ Springboot/cuidapcd/src/main/java/com/poo/cuidapcd/entity/Cliente.java
@@ -3,7 +3,7 @@ package com.poo.cuidapcd.entity;
 public class Cliente extends Usuario{
     private String preferencias;
 
-    public Cliente(int id, String nome, String email, String senha, String telefone, String cpf, 
+    public Cliente(Long id, String nome, String email, String senha, String telefone, String cpf, 
             String preferencias) {
         super(id, nome, email, senha, telefone, cpf);
         this.preferencias = preferencias;
